+++ src/lib_shell/prevalidator_internal.ml
@@ -249,9 +249,7 @@ end
 module Make_s
     (Filter : Shell_plugin.FILTER)
     (Prevalidation_t : Prevalidation.T
-                         with type validation_state =
-                           Filter.Proto.validation_state
-                          and type protocol_operation = Filter.Proto.operation) :
+                         with type protocol_operation = Filter.Proto.operation) :
   S
     with type config = Prevalidation_t.config
      and type protocol_operation = Filter.Proto.operation
@@ -344,10 +342,9 @@ module Make_s
     | `None -> `None
 
   (* This function retrieves an old/replaced operation and reclassifies it as
-     [replacement_classification]. Note that we don't need to re-flush the
-     mempool, as this function is only called in precheck mode.
+     [replacement_classification].
 
-     The operation is expected to be (a) parsable and (b) in the "prechecked"
+     The operation is expected to be (a) parsable and (b) in the "validated"
      class. So, we softly handle the situations where the operation is
      unparsable or not found in any class in case this invariant is broken
      for some reason.
@@ -395,14 +392,14 @@ module Make_s
     let to_handle = (op, classification) :: to_replace in
     let mempool =
       match classification with
-      | `Prechecked | `Applied -> Mempool.cons_valid op.hash mempool
+      | `Validated -> Mempool.cons_valid op.hash mempool
       | `Branch_refused _ | `Branch_delayed _ | `Refused _ | `Outdated _ ->
           mempool
     in
     return (v_state, mempool, to_handle)
 
-  (* Classify pending operations into either: [Refused |
-     Branch_delayed | Branch_refused | Applied | Outdated].
+  (* Classify pending operations into either:
+     [Refused | Outdated | Branch_delayed | Branch_refused | Validated].
      To ensure fairness with other worker requests, classification of
      operations is done by batch of [operation_batch_size] operations.
 
@@ -411,10 +408,8 @@ module Make_s
      - If an operation is classified, it is not part of the [pending]
      map
 
-     - A classified operation is part of the [in_mempool] set
-
-     - A classified operation is part only of one of the following
-     classes: [Branch_refused, Branch_delayed, Refused, Applied]
+     - See {!type-Prevalidator_classification.t} for additional details
+     and invariants on the classifications themselves.
 
      Moreover, this function ensures that only each newly classified
      operations are advertised to the remote peers. However, if a peer
@@ -464,21 +459,17 @@ module Make_s
       in
       advertise pv_shell mempool_to_advertise ;
       let our_mempool =
-        let prechecked_hashes =
+        let validated_hashes =
           (* Outputs hashes in "decreasing" order which should not matter *)
           Classification.Sized_map.fold
             (fun x _ acc -> x :: acc)
-            pv_shell.classification.prechecked
+            pv_shell.classification.validated
             []
         in
         {
           (* FIXME: https://gitlab.com/tezos/tezos/-/issues/2065
-             This field does not only contain valid operation *)
-          Mempool.known_valid =
-            List.fold_left
-              (fun acc op -> op.hash :: acc)
-              prechecked_hashes
-              pv_shell.classification.applied_rev;
+             This field does not need to be a list. *)
+          Mempool.known_valid = validated_hashes;
           pending = Pending_ops.hashes pv_shell.pending;
         }
       in
@@ -666,11 +657,10 @@ module Make_s
                   to_handle
               in
               match op_status with
-              | Some (_h, (`Applied | `Prechecked)) ->
+              | Some (_h, `Validated) ->
                   (* TODO: https://gitlab.com/tezos/tezos/-/issues/2294
-                     In case of `Passed_precheck_with_replace, we may want to only do
-                     the injection/replacement if a flag `replace` is set to true
-                     in the injection query. *)
+                     We may want to only do the injection/replacement if a
+                     flag `replace` is set to true in the injection query. *)
                   let*! () =
                     pv.shell.parameters.tools.chain_tools.inject_operation
                       oph
@@ -694,7 +684,7 @@ module Make_s
                     | `Outdated e ) ) ->
                   Lwt.return
                   @@ error_with
-                       "Error while applying operation %a:@ %a"
+                       "Error while validating injected operation %a:@ %a"
                        Operation_hash.pp
                        oph
                        pp_print_trace
@@ -781,13 +771,13 @@ module Make_s
               ~mempool
               shell.predecessor
 
-    (* If [flush_if_prechecked] is [true], removing a prechecked
+    (* If [flush_if_validated] is [true], removing a validated
        operation triggers a flush of the mempool. Because flushing may
        be costly this should be done only when the action is triggered
        locally by the user. This allows a better UX if the user bans a
-       prechecked operation so that a branch delayed operation becomes
-       [applied] again. *)
-    let remove ~flush_if_prechecked pv oph =
+       [validated] operation with the express goal to allow a
+       [branch_delayed] operation to become [validated] again. *)
+    let remove ~flush_if_validated pv oph =
       let open Lwt_result_syntax in
       pv.shell.parameters.tools.chain_tools.clear_or_cancel oph ;
       pv.shell.advertisement <-
@@ -800,14 +790,8 @@ module Make_s
           pv.shell.fetching <- Operation_hash.Set.remove oph pv.shell.fetching ;
           return_unit
       | Some (_op, classification) -> (
-          match (classification, flush_if_prechecked) with
-          | `Prechecked, true | `Applied, _ ->
-              (* Modifying the list of operations classified as [Applied]
-                 might change the classification of all the operations in
-                 the mempool. Hence if the removed operation has been
-                 applied we flush the mempool to force the
-                 reclassification of all the operations except the one
-                 removed. *)
+          match (classification, flush_if_validated) with
+          | `Validated, true ->
               let+ () =
                 on_flush
                   ~handle_branch_refused:false
@@ -821,7 +805,7 @@ module Make_s
           | `Branch_refused _, _
           | `Refused _, _
           | `Outdated _, _
-          | `Prechecked, false ->
+          | `Validated, false ->
               pv.validation_state <-
                 Prevalidation_t.remove_operation pv.validation_state oph ;
               return_unit)
@@ -830,7 +814,7 @@ module Make_s
       let open Lwt_result_syntax in
       pv.shell.banned_operations <-
         Operation_hash.Set.add oph_to_ban pv.shell.banned_operations ;
-      let* res = remove ~flush_if_prechecked:true pv oph_to_ban in
+      let* res = remove ~flush_if_validated:true pv oph_to_ban in
       let*! () = Events.(emit operation_banned) oph_to_ban in
       return res
   end
@@ -859,9 +843,7 @@ module Make
     (Filter : Shell_plugin.FILTER)
     (Arg : ARG)
     (Prevalidation_t : Prevalidation.T
-                         with type validation_state =
-                           Filter.Proto.validation_state
-                          and type protocol_operation = Filter.Proto.operation
+                         with type protocol_operation = Filter.Proto.operation
                           and type chain_store = Store.chain_store) : T = struct
   module S = Make_s (Filter) (Prevalidation_t)
   open S
@@ -992,22 +974,11 @@ module Make
           !dir
           (Proto_services.S.Mempool.pending_operations Tezos_rpc.Path.open_root)
           (fun pv params () ->
-            (* FIXME https://gitlab.com/tezos/tezos/-/issues/2250
-
-               We merge prechecked operation with applied operation
-               so that the encoding of the RPC does not need to be
-               changed. Once prechecking will be done by the protocol
-               and not the plugin, we will change the encoding to
-               reflect that. *)
-            let prechecked_with_applied =
-              if params#applied then
-                let applied_seq =
-                  pv.shell.classification.applied_rev |> List.to_seq
-                  |> Seq.map (fun ({hash; _} as op) -> (hash, op))
-                in
+            let validated =
+              if params#validated then
                 Classification.Sized_map.to_map
-                  pv.shell.classification.prechecked
-                |> Operation_hash.Map.to_seq |> Seq.append applied_seq
+                  pv.shell.classification.validated
+                |> Operation_hash.Map.to_seq
                 |> Seq.filter_map (fun (oph, op) ->
                        if
                          filter_validation_passes
@@ -1063,7 +1034,7 @@ module Make
             in
             let pending_operations =
               {
-                Proto_services.Mempool.applied = prechecked_with_applied;
+                Proto_services.Mempool.validated;
                 refused;
                 outdated;
                 branch_refused;
@@ -1089,20 +1060,10 @@ module Make
               Lwt_watcher.create_stream pv.operation_stream
             in
             (* First call : retrieve the current set of op from the mempool *)
-            let applied_seq =
-              if params#applied then
-                pv.shell.classification.applied_rev |> List.to_seq
-                |> Seq.map (fun {hash; protocol; _} -> ((hash, protocol), None))
-              else Seq.empty
-            in
-            (* FIXME https://gitlab.com/tezos/tezos/-/issues/2250
-
-               For the moment, applied and prechecked operations are
-               handled the same way for the user point of view. *)
-            let prechecked_seq =
-              if params#applied then
+            let validated_seq =
+              if params#validated then
                 Classification.Sized_map.to_map
-                  pv.shell.classification.prechecked
+                  pv.shell.classification.validated
                 |> Operation_hash.Map.to_seq
                 |> Seq.map (fun (hash, {protocol; _}) ->
                        ((hash, protocol), None))
@@ -1144,12 +1105,12 @@ module Make
             let current_mempool =
               Seq.append outdated_seq branch_delayed_seq
               |> Seq.append branch_refused_seq
-              |> Seq.append refused_seq |> Seq.append prechecked_seq
-              |> Seq.append applied_seq |> Seq.filter filter |> List.of_seq
+              |> Seq.append refused_seq |> Seq.append validated_seq
+              |> Seq.filter filter |> List.of_seq
             in
             let current_mempool = ref (Some current_mempool) in
             let filter_result = function
-              | `Prechecked | `Applied -> params#applied
+              | `Validated -> params#validated
               | `Refused _ -> params#refused
               | `Outdated _ -> params#outdated
               | `Branch_refused _ -> params#branch_refused
@@ -1167,7 +1128,7 @@ module Make
                   | Some (kind, op) when filter_result kind ->
                       let errors =
                         match kind with
-                        | `Prechecked | `Applied -> None
+                        | `Validated -> None
                         | `Branch_delayed errors
                         | `Branch_refused errors
                         | `Refused errors
@@ -1331,11 +1292,10 @@ module Make
           worker = mk_worker_tools w;
         }
       in
-      Shell_metrics.Mempool.set_applied_collector (fun () ->
-          List.length shell.classification.applied_rev |> float_of_int) ;
+      Shell_metrics.Mempool.set_applied_collector (fun () -> 0.) ;
       Shell_metrics.Mempool.set_prechecked_collector (fun () ->
           Prevalidator_classification.Sized_map.cardinal
-            shell.classification.prechecked
+            shell.classification.validated
           |> float_of_int) ;
       Shell_metrics.Mempool.set_refused_collector (fun () ->
           Prevalidator_classification.cardinal shell.classification.refused