+++ src/lib_shell/test/test_prevalidator_classification_operations.ml
@@ -462,8 +462,7 @@ module Recyle_operations = struct
     in
     let expected_from_classification =
       Classification.Internal_for_tests.to_map
-        ~applied:true
-        ~prechecked:true
+        ~validated:true
         ~branch_delayed:true
         ~branch_refused:handle_branch_refused
         ~refused:false
@@ -513,8 +512,7 @@ module Recyle_operations = struct
     in
     let expected : unit operation Op_map.t =
       Classification.Internal_for_tests.to_map
-        ~applied:false
-        ~prechecked:false
+        ~validated:false
         ~branch_delayed:false
         ~branch_refused:(not handle_branch_refused)
         ~refused:true
@@ -538,8 +536,7 @@ module Recyle_operations = struct
     in
     let actual =
       Classification.Internal_for_tests.to_map
-        ~applied:true
-        ~prechecked:true
+        ~validated:true
         ~branch_delayed:true
         ~branch_refused:true
         ~refused:true