+++ CHANGES.rst
@@ -127,6 +127,15 @@ Node
   ``double_attestation_evidence`` and ``double_preattestation_evidence`` kinds
   in the JSON result. (MR :gl:`!8949`)
 
+- The error message when the local injection of an operation fails now
+  begins with ``Error while validating injected operation`` instead of
+  ``Error while applying operation``. (MR :gl:`!8857`)
+
+- Updated the description of the ``ban_operation`` RPC to better
+  reflect its behavior, which is unchanged. (More precisely, removed
+  the "reverting its effect if it was applied" part since operations
+  are never applied.) (MR :gl:`!8857`)
+
 Client
 ------
 - Adding client commands to generate, open and verify a time-lock.